Minor changes to use create_query

+++ tests/C/query/common/adios_query_xml_parse.c
@@ -140,8 +140,8 @@ ADIOS_QUERY_TEST_INFO * parseXml(const char *inputxml, ADIOS_FILE* f) {
 		size_t bytes_read = fread (buffer, 1, s.st_size, fp);
 
 		if (bytes_read != s.st_size) {
-			fprintf(stderr, "error reading input xml file: %s. Expected %ld Got %lld\n"
-					,inputxml, s.st_size, (long long int)bytes_read );
+			fprintf(stderr, "error reading input xml file: %s. Expected %ld Got %ld\n"
+					,inputxml, s.st_size, bytes_read );
 			fclose(fp);
 			return NULL;
 		}
